Added Data Validations

+++ api/Controllers/StockController.cs
@@ -31,15 +31,21 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var stocks = await _stockRepository.GetAllAsync();
             var stocksDto = stocks.Select(stock => stock.ToStockDto());
             return Ok(stocksDto);
         }
 
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute]int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepository.GetByIdAsync(id);
                 
             if (stock == null)
@@ -53,6 +59,9 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody]CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDto();
 
             await _stockRepository.CreateAsync(stockModel);
@@ -60,9 +69,12 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute]int id, [FromBody]UpdateStockRequestDto updateStockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepository.UpdateAsync(id, updateStockDto);
             if (stockModel == null)
             {
@@ -72,9 +84,12 @@ namespace api.Controllers
             return Ok(stockModel.ToStockDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepository.DeleteAsync(id);
             if (stockModel == null)
             {
